refactor(ScannerRun)!: Store scan results by provenance
Introduce an `OrtResult` internal format for storing the scan results
by provenance, but map it back in `getScanResults()` and
`getScanResultsById()` to the previously used representation, to
decouple the calling code from the internal representation and to hide
its complexities. This eliminates redundancy in case multiple packages
share the same provenance.

This allows to add file lists with sha1 sums of all scanned file, also
stored "by provenance" in a consistent way in an upcomining change.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>